refactor: format java-sdk code (#210)
