+++ sdk/java-sdk/src/main/java/spec/sdk/reactor/v1/CoreCloudRuntimes.java
@@ -1,11 +1,9 @@
 package spec.sdk.reactor.v1;
 
-import spec.sdk.reactor.v1.domain.core.BindingRuntimes;
 import spec.sdk.reactor.v1.domain.core.ConfigurationRuntimes;
 import spec.sdk.reactor.v1.domain.core.InvocationRuntimes;
 import spec.sdk.reactor.v1.domain.core.PubSubRuntimes;
 import spec.sdk.reactor.v1.domain.core.StateRuntimes;
-import spec.sdk.reactor.v1.domain.core.SecretsRuntimes;
 
 /**
  * Core Cloud Runtimes standard API defined.
@@ -13,8 +11,8 @@ import spec.sdk.reactor.v1.domain.core.SecretsRuntimes;
 public interface CoreCloudRuntimes extends
         InvocationRuntimes,
         PubSubRuntimes,
-        //BindingRuntimes,
+        // BindingRuntimes,
         StateRuntimes,
-        //SecretsRuntimes,
+        // SecretsRuntimes,
         ConfigurationRuntimes {
 }
