--- sdk/java-sdk/src/main/java/spec/sdk/runtime/v1/RPCClient.java
@@ -1,4 +1,0 @@
-package spec.sdk.runtime.v1;
-
-public interface RPCClient {
-}
