--- sdk/java-sdk/src/main/java/spec/sdk/runtime/v1/RuntimeClient.java
@@ -1,5 +1,0 @@
-package spec.sdk.runtime.v1;
-
-public interface RuntimeClient extends HelloClient,
-        RPCClient, PubSubClient, StateClient {
-}
