+++ sdk/java-sdk/src/main/java/spec/sdk/runtime/v1/client/RuntimesClient.java
@@ -1,0 +1,11 @@
+package spec.sdk.runtime.v1.client;
+
+import spec.sdk.runtime.v1.domain.*;
+
+public interface RuntimesClient extends
+        HelloRuntimes,
+        ConfigurationRuntimes,
+        InvocationRuntimes,
+        PubSubRuntimes,
+        StateRuntimes {
+}
