+++ sdk/java-sdk/src/main/java/spec/sdk/runtime/v1/domain/InvocationRuntimes.java
@@ -1,0 +1,4 @@
+package spec.sdk.runtime.v1.domain;
+
+public interface InvocationRuntimes {
+}
